refactor(blocks-construct): code refactor in file manager feature
