+++ src/features/file-manager/constants/file-manager.constant.ts
@@ -1,0 +1,9 @@
+export const FILE_BREADCRUMB_TITLES = {
+  '/files': 'MY_FILES',
+  '/files/1': 'MEETING_NOTES',
+  '/files/2': 'RESEARCH_DATA',
+  '/files/3': 'CLIENT_DOCUMENTS',
+  '/files/4': 'PROJECT_FILES',
+  '/files/5': 'DESIGN_ASSETS',
+  '/files/11': 'MARKETING_ASSETS',
+};
