+++ src/features/file-manager/hooks/use-file-manager.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable no-console */
 import { useCallback, useState } from 'react';
-import { IFileDataWithSharing, SharedUser, UserRole } from '../utils/file-manager';
 import { v4 as uuidv4 } from 'uuid';
+import { IFileDataWithSharing, SharedUser, UserRole } from '../utils/file-manager';
 
 interface UseFileManagerProps {
   onCreateFile?: () => void;
