+++ src/features/file-manager/types/file-manager.type.ts
@@ -1,46 +1,35 @@
-import { IFileData } from '../hooks/use-mock-files-query';
-import { IFileDataWithSharing, SharedUser } from '../utils/file-manager';
-import { SharedFilters } from './header-toolbar.type';
-
 export interface DateRange {
   from?: Date;
   to?: Date;
 }
+export type FileType = 'Folder' | 'File' | 'Image' | 'Audio' | 'Video';
 
-export interface SharedFilesListViewProps {
-  onDownload: ((file: IFileData) => void) | undefined;
-  onViewDetails: (file: IFileDataWithSharing) => void;
-  onShare: (file: IFileDataWithSharing) => void;
-  onDelete: (file: IFileDataWithSharing) => void;
-  onMove: (file: IFileDataWithSharing) => void;
-  onCopy: (file: IFileDataWithSharing) => void;
-  onOpen: (file: IFileDataWithSharing) => void;
-  onRename: (file: IFileDataWithSharing) => void;
-  filters: SharedFilters;
-  newFiles?: IFileDataWithSharing[];
-  newFolders?: IFileDataWithSharing[];
-  renamedFiles?: Map<string, IFileDataWithSharing>;
-  fileSharedUsers?: { [key: string]: SharedUser[] };
-  filePermissions?: { [key: string]: { [key: string]: string } };
-  currentFolderId?: string;
-  onNavigateToFolder?: (folderId: string) => void;
+export interface IFileData {
+  id: string;
+  name: string;
+  lastModified: Date;
+  fileType: FileType;
+  size: string;
+  isShared?: boolean;
+  sharedBy?: {
+    id: string;
+    name: string;
+    avatar?: string;
+  };
+  sharedDate?: Date;
+  parentFolderId?: string;
 }
 
-export interface SharedFilesGridViewProps {
-  onViewDetails?: (file: IFileDataWithSharing) => void;
-  onFilePreview?: (file: IFileDataWithSharing) => void;
-  onShare: (file: IFileDataWithSharing) => void;
-  onDelete: (file: IFileDataWithSharing) => void;
-  onMove: (file: IFileDataWithSharing) => void;
-  onCopy: (file: IFileDataWithSharing) => void;
-  onRename: (file: IFileDataWithSharing) => void;
-  filters: any;
-  newFiles?: IFileDataWithSharing[];
-  newFolders?: IFileDataWithSharing[];
-  renamedFiles?: Map<string, IFileDataWithSharing>;
-  fileSharedUsers?: { [key: string]: SharedUser[] };
-  filePermissions?: { [key: string]: { [key: string]: string } };
-  currentFolderId?: string;
-  onNavigateToFolder?: (folderId: string) => void;
-  onNavigateBack?: () => void;
+export interface QueryParams {
+  filter: {
+    name?: string;
+    fileType?: FileType;
+    lastModified?: {
+      from?: Date;
+      to?: Date;
+    };
+  };
+  page: number;
+  pageSize: number;
+  folderId?: string;
 }
